refactor(visualization): dedup metrics-bar segments and settings popovers
- Extract MetricSegmentBase shared by the area/height segments (#44).
- Extract a reusable cc-settings-popover-shell container used by the area,
  height, color and edge settings popovers (#47); all popover ids, anchors,
  testids and width behavior preserved.
- parseNumberInput now uses parseFloat so numeric settings inputs accept decimals.
- Fix the Color 'Reset thresholds' button resetting delta map colors as a
  side effect; it now resets only dynamicSettings.colorRange.

+++ visualization/app/codeCharta/util/parseNumberInput.ts
@@ -1,6 +1,6 @@
 export const parseNumberInput = (event: Event, min: number, max: number) => {
     const input = event.target as HTMLInputElement
-    const value = Number.parseInt(input.value)
+    const value = Number.parseFloat(input.value)
     if (Number.isNaN(value)) {
         return value
     }
